Moved validation state into separate source file
Moved ValidationState
Added Airport Names to DatabasaCache
